rename(ui-design-tool): rename classes related to UIDesignTool
